Refactor MCP tools into domain modules and shared helpers.
Split the God tools.ts, dedupe metadata filter definitions, centralize MCP user-id extraction, comment notify fan-out, and Result→HTTP error mapping so REST and MCP stay aligned without copy-paste.
