+++ web/app/api/v1/instruments/[instrumentId]/runs/[runId]/reprocess/route.ts
@@ -1,6 +1,6 @@
 import type { NextRequest } from "next/server";
 import { authorize } from "@/lib/api/auth";
-import { apiError, CONFLICT, NOT_FOUND } from "@/lib/api/errors";
+import { apiErrorFromResult } from "@/lib/api/errors";
 import { reprocessRun } from "@/lib/api/file-reprocessing";
 
 interface RouteContext {
@@ -24,8 +24,7 @@ export async function POST(request: NextRequest, { params }: RouteContext) {
   const result = await reprocessRun(instrumentId, runId);
 
   if (!result.ok) {
-    const code = result.code === "NOT_FOUND" ? NOT_FOUND : CONFLICT;
-    return apiError(result.status, code, result.message);
+    return apiErrorFromResult(result);
   }
 
   return Response.json({
