+++ web/lib/mcp/validate-run-filters.ts
@@ -1,102 +1,34 @@
-import {
-  getInstrumentFilterOptions,
-  type InstrumentFilterOptionsByType,
-} from "@/lib/api/instrument-runs";
+import { getInstrumentFilterOptions } from "@/lib/api/instrument-runs";
 import { getInstrumentById } from "@/lib/api/instruments";
+import {
+  RUN_METADATA_FILTER_DEFS,
+  type RunMetadataFilterArgs,
+  type RunMetadataFilterKey,
+} from "@/lib/api/run-metadata-filters";
 
-interface MetadataFilterArgs {
-  captureType?: string;
-  colorMode?: string;
-  dpi?: string;
-  dyeChannel?: string;
-  gelColor?: string;
-  gelWavelength?: string;
-  hinaChannel?: string;
-  hinaDimension?: string;
-  hinaSize?: string;
-  imagingMode?: string;
-  measurementMode?: string;
-  measurementType?: string;
-  wavelength?: string;
-}
+const FILTER_KEYS = RUN_METADATA_FILTER_DEFS.map(
+  (def) => def.key
+) as RunMetadataFilterKey[];
 
-// The instrument kind each metadata filter belongs to. A filter is only
-// applicable when the scoped instrument's kind matches; supplying one that
-// doesn't (e.g. a plate-reader `wavelength` on a gel-doc) is rejected below.
-const KEY_TO_KIND = {
-  wavelength: "plate_reader",
-  measurementMode: "plate_reader",
-  measurementType: "plate_reader",
-  captureType: "gel_doc",
-  imagingMode: "gel_doc",
-  gelWavelength: "gel_doc",
-  gelColor: "gel_doc",
-  dyeChannel: "qpcr",
-  hinaChannel: "hina_microscope",
-  hinaDimension: "hina_microscope",
-  hinaSize: "hina_microscope",
-  dpi: "epson_v700_scanner",
-  colorMode: "epson_v700_scanner",
-} as const satisfies Record<
-  keyof MetadataFilterArgs,
-  Exclude<InstrumentFilterOptionsByType["kind"], "default">
+const KEY_TO_KIND = Object.fromEntries(
+  RUN_METADATA_FILTER_DEFS.map((def) => [def.key, def.kind])
+) as Record<
+  RunMetadataFilterKey,
+  (typeof RUN_METADATA_FILTER_DEFS)[number]["kind"]
 >;
 
-// Allowed values for a key on an instrument whose kind already matches the
-// key (guaranteed by the `KEY_TO_KIND` check in the caller).
-function allowedList(
-  options: InstrumentFilterOptionsByType,
-  key: keyof MetadataFilterArgs
-): string[] {
-  switch (key) {
-    case "wavelength":
-      return options.kind === "plate_reader" ? options.options.wavelengths : [];
-    case "measurementMode":
-      return options.kind === "plate_reader"
-        ? options.options.measurementModes
-        : [];
-    case "measurementType":
-      return options.kind === "plate_reader"
-        ? options.options.measurementTypes
-        : [];
-    case "captureType":
-      return options.kind === "gel_doc" ? options.options.captureTypes : [];
-    case "imagingMode":
-      return options.kind === "gel_doc" ? options.options.imagingModes : [];
-    case "gelWavelength":
-      return options.kind === "gel_doc" ? options.options.wavelengths : [];
-    case "gelColor":
-      return options.kind === "gel_doc" ? options.options.colors : [];
-    case "dyeChannel":
-      return options.kind === "qpcr" ? options.options.dyeChannels : [];
-    case "hinaChannel":
-      return options.kind === "hina_microscope" ? options.options.channels : [];
-    case "hinaDimension":
-      return options.kind === "hina_microscope"
-        ? options.options.dimensions
-        : [];
-    case "hinaSize":
-      return options.kind === "hina_microscope"
-        ? options.options.sizes.map((s) => s.value)
-        : [];
-    case "dpi":
-      return options.kind === "epson_v700_scanner" ? options.options.dpis : [];
-    case "colorMode":
-      return options.kind === "epson_v700_scanner"
-        ? options.options.colorModes
-        : [];
-    default:
-      return [];
-  }
-}
-
-const FILTER_KEYS = Object.keys(KEY_TO_KIND) as (keyof MetadataFilterArgs)[];
+const KEY_TO_ALLOWED = Object.fromEntries(
+  RUN_METADATA_FILTER_DEFS.map((def) => [def.key, def.allowedValues])
+) as Record<
+  RunMetadataFilterKey,
+  (typeof RUN_METADATA_FILTER_DEFS)[number]["allowedValues"]
+>;
 
 // When a single instrument is scoped, reject metadata values that are not in
 // that instrument's filter-options so agents get a corrective enum list.
 export async function validateSearchRunsMetadataFilters(
   instrumentId: string,
-  args: MetadataFilterArgs
+  args: RunMetadataFilterArgs
 ): Promise<string | null> {
   const hasAny = FILTER_KEYS.some((key) => args[key] !== undefined);
   if (!hasAny) {
@@ -132,7 +64,7 @@ export async function validateSearchRunsMetadataFilters(
         `See datahub://instruments/${instrumentId}/filter-options for applicable filters.`
       );
     }
-    const allowed = allowedList(options, key);
+    const allowed = KEY_TO_ALLOWED[key](options);
     if (!allowed.includes(value)) {
       return (
         `Invalid ${key}="${value}" for instrument '${instrumentId}'. ` +
